# backend/docs/openapi.yaml
definitions:
  auth.TokenPair:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        description: Creation timestamp (UTC)
        format: date-time
        type: string
      email:
        description: Email address (unique)
        example: user@example.com
        format: email
        type: string
      email_verified:
        description: Whether the email has been verified
        example: false
        type: boolean
      founder_id:
        description: Related founder profile ID
        example: 1
        type: integer
      id:
        description: Unique user identifier
        example: 1
        type: integer
      image_url:
        description: Avatar URL
        example: https://cdn.example.com/avatars/1.png
        format: uri
        type: string
      investor_id:
        description: Related investor profile ID
        example: 2
        type: integer
      name:
        description: Display name
        example: Jane Doe
        type: string
      role:
        description: Role name
        enum:
        - admin
        - user
        - investor
        - founder
        example: user
        type: string
      updated_at:
        description: Update timestamp (UTC)
        format: date-time
        type: string
    type: object
  requests.AuthForgotPasswordRequest:
    properties:
      email:
        description: Account email to send reset instructions
        example: john@doe.tld
        format: email
        type: string
    type: object
  requests.AuthLoginRequest:
    properties:
      email:
        description: Email address
        example: john@doe.tld
        format: email
        type: string
      password:
        description: Password
        example: secret123
        type: string
    type: object
  requests.AuthRefreshRequest:
    properties:
      refresh_token:
        description: Valid refresh token
        example: <jwt>
        type: string
    type: object
  requests.AuthRegisterRequest:
    properties:
      email:
        description: Email address of the user
        example: john@doe.tld
        format: email
        type: string
      founder_id:
        description: Optional associated founder profile ID
        example: 1
        type: integer
      image_url:
        description: Optional profile image URL
        example: https://cdn.example.com/avatars/john.png
        format: uri
        type: string
      investor_id:
        description: Optional associated investor profile ID
        example: 2
        type: integer
      name:
        description: Full name to display
        example: John Doe
        type: string
      password:
        description: Plain password (min 6 chars)
        example: secret123
        type: string
      role:
        description: Account role
        enum:
        - investor
        - founder
        example: founder
        type: string
    type: object
  requests.AuthResetPasswordRequest:
    properties:
      new_password:
        description: New password to set
        example: secret123
        type: string
      token:
        description: Password reset token
        example: <reset-token>
        type: string
    type: object
  requests.AuthVerifyRequest:
    properties:
      token:
        description: One-time verification token
        example: <verify-token>
        type: string
    type: object
...
  response.AuthRegisterResponse:
    properties:
      tokens:
        $ref: '#/definitions/auth.TokenPair'
      user:
        $ref: '#/definitions/models.User'
    type: object
  response.ErrorBody:
    properties:
      code:
        description: Machine-readable error code
        enum:
        - invalid_params
        - invalid_sort
        - invalid_payload
        - internal_error
        - not_found
        - email_taken
        - invalid_token
        - invalid_credentials
        - no_fields
        - method_not_allowed
        - unauthorized
        example: internal_error
        type: string
      details:
        description: Optional granular details (validation errors, etc.)
      message:
        description: Human-readable error message
        example: failed to retrieve resource
        type: string
    type: object
...
  response.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  response.TokenPairResponse:
    properties:
      tokens:
        $ref: '#/definitions/auth.TokenPair'
    type: object
...
info:
  contact: {}
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a password reset email (if the user exists).
      parameters:
      - description: Email
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.AuthForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
      summary: Forgot password
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Verifies credentials and returns a JWT token pair.
      parameters:
      - description: Credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.AuthLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AuthRegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      summary: Sign in
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a valid refresh token for a new token pair.
      parameters:
      - description: Refresh token
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.AuthRefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TokenPairResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      summary: Refresh tokens
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a user and returns the profile plus JWT tokens.
      parameters:
      - description: Registration data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.AuthRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.AuthRegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      summary: Sign up
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Updates the password using a valid reset token.
      parameters:
      - description: New password
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.AuthResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      summary: Reset password
      tags:
      - Auth
  /auth/verify:
    get:
      description: Verifies the user's email using a one-time token.
      parameters:
      - description: Verification token
        in: query
        name: token
        type: string
      - description: Token in body
        in: body
        name: payload
        schema:
          $ref: '#/definitions/requests.AuthVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      summary: Verify email
      tags:
      - Auth
    post:
      description: Verifies the user's email using a one-time token.
      parameters:
      - description: Verification token
        in: query
        name: token
        type: string
      - description: Token in body
        in: body
        name: payload
        schema:
          $ref: '#/definitions/requests.AuthVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      summary: Verify email
      tags:
      - Auth
...
swagger: "2.0"